Separated into own assembly definition.
