+++ src/MessagePack.UnityClient/Assets/Scripts/MessagePack/Unity/Mathematics.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d694dfc89ae6776479694c18441bd1a2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
