+++ src/MessagePack.UnityClient/Assets/Scripts/MessagePack/Unity/Mathematics/MessagePack.Mathematics.asmdef
@@ -1,0 +1,25 @@
+{
+    "name": "MessagePack.Mathematics",
+    "references": [
+        "MessagePack",
+        "Unity.Mathematics"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": true,
+    "precompiledReferences": [
+        "System.Buffers.dll",
+        "System.Memory.dll"
+    ],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [
+        {
+            "name": "com.unity.mathematics",
+            "expression": "",
+            "define": "USE_UNITY_MATHEMATICS"
+        }
+    ],
+    "noEngineReferences": true
+}
