+++ src/MessagePack.UnityClient/Assets/Scripts/MessagePack/Unity/Mathematics/MessagePack.Mathematics.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: cce79b9eb6c7e8947b550d99060db470
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
