+++ src/MessagePack.UnityClient/Assets/Scripts/MessagePack/Unity/Mathematics/UnityMathematicsFormatters.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6777bb03d10ef4d48b92c613f3a8f5fc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
