+++ src/MessagePack.UnityClient/Assets/Scripts/MessagePack/Unity/Mathematics/UnityMathematicsResolver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 09bb433f09f346247ac94204baecc7db
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
